refactor(web): relocate library code to domains/library/, inline callbacks, remove dead code (LUM-2007)
Move library-specific files from components/apps/ (shared UI directory) to
domains/library/ where they belong per CONVENTIONS.md. Delete 3 unnecessary
hooks (useAppViewer, useAppDelete, useAppImport) that were just useState +
useCallback wrappers without data-fetching composition. Inline their
callbacks in the orchestrator per the established domains/home/ pattern.

Remove dead code:
- Inline app viewer branch (unreachable: only consumer always provides onOpenApp)
- onEditApp prop (never passed by any consumer)
- isOpening/openingAppId state (navigation is synchronous)
- document-viewer-container.tsx (zero importers)

Architecture now follows CONVENTIONS.md:
- domains/library/hooks/ — data-fetching hook (TanStack Query)
- domains/library/components/ — responsibility-based sub-components
- domains/library/utils/ — pure utilities
- domains/library/library-view.tsx — thin orchestrator (~265 lines)

Closes LUM-2007
