+++ apps/web/src/domains/library/components/library-empty-state.tsx
