+++ apps/web/src/domains/library/hooks/use-library-data.ts
@@ -67,7 +67,6 @@ export function useLibraryData(assistantId: string) {
     pinnedApps,
     recentApps,
     filteredDocuments,
-    pinnedAppIds,
     searchText,
     setSearchText,
     loading,
